Updated the material color unknowns to floats
Still have no idea what these fields are for, but they are clearly all
floating point 1.0

+++ src/libs/jkcontent/material.cpp
@@ -18,7 +18,7 @@ gorc::material_cel_record::material_cel_record(deserialization_constructor_tag,
 
     color_index = binary_deserialize<uint32_t>(bis);
     for(auto &em : unknown_color) {
-        em = binary_deserialize<uint32_t>(bis);
+        em = binary_deserialize<float>(bis);
     }
 
     if(type == 8) {
